# Improve `ActiveSupport::Duration` performance about 25% ~ 50%