`ActiveSupport::Duration` expects `parts` as a Hash in the internal, but
sometimes that is passed as nested array.

(e.g. `parts.map { |type, number| [type, number * other] }`)

By avoiding the extra array allocation, it makes
`ActiveSupport::Duration` performance faster about 25% ~ 50%.

```ruby
Benchmark.ips do |x|
  x.report("4.seconds + 2.seconds") { 4.seconds + 2.seconds }
  x.report("4.seconds - 2.seconds") { 4.seconds - 2.seconds }
  x.report("4.seconds * 2") { 4.seconds * 2 }
  x.report("4.seconds / 2") { 4.seconds / 2 }
end
```

Before (e36d4a03):

```
Warming up --------------------------------------
4.seconds + 2.seconds   27.515k i/100ms
4.seconds - 2.seconds   20.783k i/100ms
       4.seconds * 2    40.286k i/100ms
       4.seconds / 2    41.717k i/100ms
Calculating -------------------------------------
4.seconds + 2.seconds   299.133k (±13.4%) i/s -      1.486M in   5.053953s
4.seconds - 2.seconds   222.476k (± 9.8%) i/s -      1.122M in   5.088532s
       4.seconds * 2    479.650k (±10.5%) i/s -      2.377M in   5.008644s
       4.seconds / 2    473.768k (±11.3%) i/s -      2.378M in   5.082413s
```

After (this change):

```
Warming up --------------------------------------
4.seconds + 2.seconds   35.856k i/100ms
4.seconds - 2.seconds   28.657k i/100ms
       4.seconds * 2    51.744k i/100ms
       4.seconds / 2    51.417k i/100ms
Calculating -------------------------------------
4.seconds + 2.seconds   445.972k (±21.6%) i/s -      2.080M in   5.018570s
4.seconds - 2.seconds   317.972k (±18.2%) i/s -      1.519M in   5.006564s
       4.seconds * 2    608.695k (±12.8%) i/s -      3.001M in   5.014321s
       4.seconds / 2    611.488k (±12.0%) i/s -      3.034M in   5.035211s
```